Moves struct iterators to iterators mod, adds comments
